requirements.txt
langchain==0.2.2
openai
google
langchain-google-genai
langchain_community
python-dotenv
streamlit
